Conflicts:
	ArduCopter/APM_Config.h
	ArduCopter/ArduCopter.pde
	ArduPlane/ArduPlane.pde
	ArduPlane/system.pde
	README.md

+++ libraries/AP_GPS/GPS.h
@@ -216,7 +216,7 @@ protected:
     bool _have_raw_velocity;
 
 	// detected baudrate
-	uint16_t _baudrate;
+	uint32_t _baudrate;
 
 private:
